### 2020-05-22 #2228 @whedon
[PRE REVIEW]: rgee: An R package for interacting with Google Earth Engine
**Submitting author:** @csaybar (<a href="http://orcid.org/0000-0003-2745-9535">Cesar Luis Aybar Camacho</a>)
**Repository:** <a href="https://github.com/r-spatial/rgee" target ="_blank">https://github.com/r-spatial/rgee</a>
**Version:** v.0.5.0
