+++ R/ee_clean.R
@@ -18,7 +18,9 @@
 #' }
 #' @export
 ee_clean_credentials <- function(email='not_defined', quiet=FALSE) {
-  ee_path <- path.expand("~/.config/earthengine")
+  oauth_func_path <- system.file("python/ee_utils.py", package = "rgee")
+  utils_py <- ee_source_python(oauth_func_path)
+  ee_path <- utils_py$ee_path()
   email_clean <- gsub("@gmail.com", "", email)
 
   if (email == 'not_defined') {
