+++ R/ee_Initialize.R
@@ -113,7 +113,7 @@ ee_Initialize <- function(email = NULL,
 
   # create a user's folder
   email_clean <- gsub("@gmail.com", "", email)
-  ee_path <- path.expand("~/.config/earthengine")
+  ee_path <- ee_utils$ee_path()
   ee_path_user <- sprintf("%s/%s", ee_path, email_clean)
   dir.create(ee_path_user, showWarnings = FALSE, recursive = TRUE)
 
@@ -234,7 +234,7 @@ ee_create_credentials_earthengine <- function(email_clean) {
   utils_py <- ee_source_python(oauth_func_path)
 
   # first step
-  ee_path <- path.expand("~/.config/earthengine")
+  ee_path <- utils_py$ee_path()
   main_ee_credential <- sprintf("%s/credentials", ee_path)
   user_ee_credential <- sprintf(
     "%s/%s/credentials",
@@ -273,7 +273,9 @@ ee_create_credentials_drive <- function(email) {
     )
   }
   # setting drive folder
-  ee_path <- path.expand("~/.config/earthengine")
+  oauth_func_path <- system.file("python/ee_utils.py", package = "rgee")
+  utils_py <- ee_source_python(oauth_func_path)
+  ee_path <- utils_py$ee_path()
   email_clean <- gsub("@gmail.com", "", email)
   ee_path_user <- sprintf("%s/%s", ee_path, email_clean)
   # drive_credentials
@@ -327,8 +329,10 @@ ee_create_credentials_gcs <- function(email) {
       call. = FALSE
     )
   }
+  oauth_func_path <- system.file("python/ee_utils.py", package = "rgee")
+  utils_py <- ee_source_python(oauth_func_path)
+  ee_path <- utils_py$ee_path()
   # setting gcs folder
-  ee_path <- path.expand("~/.config/earthengine")
   email_clean <- gsub("@gmail.com", "", email)
   ee_path_user <- sprintf("%s/%s", ee_path, email_clean)
   # gcs_credentials
@@ -373,8 +377,11 @@ ee_users <- function() {
   wsc <- "     "
   title  <- c('user', ' EE', ' GD', ' GCS')
 
+  oauth_func_path <- system.file("python/ee_utils.py", package = "rgee")
+  utils_py <- ee_source_python(oauth_func_path)
+
   # get all dirfiles
-  ee_path <- path.expand("~/.config/earthengine") %>%
+  ee_path <- utils_py$ee_path() %>%
     list.dirs(full.names = FALSE) %>%
     '['(-1)
 
@@ -460,9 +467,11 @@ ee_sessioninfo <- function(email = NULL,
                            user = NULL,
                            drive_cre = NULL,
                            gcs_cre = NULL) {
+  oauth_func_path <- system.file("python/ee_utils.py", package = "rgee")
+  utils_py <- ee_source_python(oauth_func_path)
   sessioninfo <- sprintf(
     "%s/rgee_sessioninfo.txt",
-    path.expand("~/.config/earthengine")
+    utils_py$ee_path()
   )
   df <- data.frame(
     email = email, user = user, drive_cre = drive_cre, gcs_cre = gcs_cre,
@@ -475,10 +484,13 @@ ee_sessioninfo <- function(email = NULL,
 #' Get the path where the credentials are stored
 #' @export
 ee_get_earthengine_path <- function() {
-  ee_path <- path.expand("~/.config/earthengine")
+  oauth_func_path <- system.file("python/ee_utils.py", package = "rgee")
+  utils_py <- ee_source_python(oauth_func_path)
+  ee_path <- utils_py$ee_path()
+
   sessioninfo <- sprintf(
[…]
