### 2020-05-24 57ec0a0
changes in ee_install
remove a Python env before to create a new one
