+++ NAMESPACE
@@ -37,6 +37,7 @@ export(ee_install)
 export(ee_install_discover_pyenvs)
 export(ee_install_earthengine_upgrade)
 export(ee_install_python_packages)
+export(ee_install_set_pyenv)
 export(ee_manage_asset_access)
 export(ee_manage_asset_size)
 export(ee_manage_assetlist)
@@ -108,6 +109,7 @@ importFrom(methods,setGeneric)
 importFrom(methods,setMethod)
 importFrom(reticulate,conda_create)
 importFrom(reticulate,conda_list)
+importFrom(reticulate,conda_remove)
 importFrom(reticulate,import)
 importFrom(reticulate,import_from_path)
 importFrom(reticulate,install_miniconda)
@@ -120,6 +122,7 @@ importFrom(reticulate,py_to_r)
 importFrom(reticulate,source_python)
 importFrom(reticulate,virtualenv_create)
 importFrom(reticulate,virtualenv_list)
+importFrom(reticulate,virtualenv_remove)
 importFrom(reticulate,virtualenv_root)
 importFrom(sf,"st_crs<-")
 importFrom(sf,"st_geometry<-")
