### 2020-05-24 17a5e5f
ee_path() now return an R character
