+++ R/ee_check.R
@@ -132,7 +132,7 @@ ee_check_rgee_python_packages <- function(quiet = FALSE) {
 ee_check_credentials <- function() {
   oauth_func_path <- system.file("python/ee_utils.py", package = "rgee")
   utils_py <- ee_source_python(oauth_func_path)
-  driverdir <- utils_py$ee_path()
+  driverdir <- ee_utils_py_to_r(utils_py$ee_path())
   ee_credentials <- sprintf("%s/credentials", driverdir)
   drive_credentials <- list.files(driverdir, "@gmail.com", full.names = TRUE)[1]
 
