### 2020-05-29 #2228 @hugoledoux
👋 @jhollist would you be interested in reviewing this submission for JOSS?

  > rgee: An R package for interacting with Google Earth Engine

software: https://github.com/r-spatial/rgee
