### 2020-05-29 #2272 @whedon
[REVIEW]: rgee: An R package for interacting with Google Earth Engine
**Submitting author:** @csaybar (<a href="http://orcid.org/0000-0003-2745-9535">Cesar Luis Aybar Camacho</a>)
**Repository:** <a href="https://github.com/r-spatial/rgee" target ="_blank">https://github.com/r-spatial/rgee</a>
**Version:** v1.0.0
