### 2020-05-29 9d7fcbf
rgee v0.5.2 release
