+++ R/ee_search.R
@@ -19,9 +19,9 @@
 #' @param path_dataset Path of the dataset. By default it will loaded
 #' automatically.
 #' @name ee_search-tools
-#' @return a data.frame.
+#' @return A data.frame where rows represents public data archive.
 #' @examples
-#' \dontrun{
+#' \donttest{
 #' library(rgee)
 #'
 #' ee_reattach() # reattach ee as a reserved word
