+++ pkgdown.yml
@@ -5,7 +5,7 @@ articles:
   considerations: considerations.html
   faq: faq.html
   setup: setup.html
-last_built: 2020-05-24T20:58Z
+last_built: 2020-05-30T10:05Z
 urls:
   reference: http://r-spatial.github.io/rgee/reference
   article: http://r-spatial.github.io/rgee/articles
