+++ vignettes/considerations.md
@@ -1,13 +1,17 @@
 `rgee` and the [Earth Engine Python
-API](https://pypi.org/project/earthengine-api/) share the same classes, functions, and methods. Therefore, is expected that works identical to all user requests. However, differences in the language design of R and Python might cause **three** bug-potential scenarios. Each of them are explained in-depth below.
+API](https://pypi.org/project/earthengine-api/) share the same classes,
+functions, and methods. Therefore, is expected that works identical to
+all user requests. However, differences in the language design of R and
+Python might cause **three** bug-potential scenarios. Each of them are
+explained in-depth below.
 
 ### 1) The **map** message error:
 
 `rgee` may show an error message when valid requests to the Earth Engine
 server are made. This issue happens when the **map** method is used
 under the next two scenarios: (1) users employing a reticulate version
-lower than &lt; 1.14 (please update it!); and (2) if you are leading
-with **ee$List** objects. For instance:
+lower than \< 1.14 (please update it!); and (2) if you are leading with
+**ee$List** objects. For instance:
 
 ``` r
 library(rgee)
@@ -36,8 +40,9 @@ mylist$map(function(x) ee$Number(x)$add(1))
 ```
 
 The code before is perfectly valid but `rgee` will produce an error.
-This problem should be easily solved by adding the function **ee_utils_pyfunc**.
-It will permit users wrap R functions before to send it to `reticulate`. Let’s see:
+This problem should be easily solved by adding the function
+**ee\_utils\_pyfunc**. It will permit users wrap R functions before to
+send it to `reticulate`. Let’s see:
 
 ``` r
 library(rgee)
@@ -94,8 +99,8 @@ and_bitwise$getInfo()
 Users need to take into consideration that most of the arguments of the
 Earth Engine methods are strict to admit only **integer values**. The
 creation of integers in R is quite simple, you just need to add the
-letter **L** to the end of the specific number or employ the
-function `as.integer`. The **correct code** in R would be:
+letter **L** to the end of the specific number or employ the function
+`as.integer`. The **correct code** in R would be:
 
 ``` r
 and_bitwise = ee$Number(32L)$bitwiseAnd(100L)
