+++ pkgdown.yml
@@ -2,10 +2,10 @@ pandoc: 2.7.3
 pkgdown: 1.5.1
 pkgdown_sha: ~
 articles:
+  BestPractices: BestPractices.html
   considerations: considerations.html
-  faq: faq.html
   setup: setup.html
-last_built: 2020-05-30T10:05Z
+last_built: 2020-06-01T00:59Z
 urls:
   reference: http://r-spatial.github.io/rgee/reference
   article: http://r-spatial.github.io/rgee/articles
