### 2020-06-01 1719455
static added to .gitignore
