### 2020-06-05 e988f1a
rgee v0.5.4 release
- Earth Engine Python API updated to 0.1.224.
- ee is now an internal rgee environment.
- ee_reattach was deleted.
- ee_print now display the ee$Image properties: system:id, system:time_start and system:time_end.
- rgee now asks users if they would like to save EARTHENGINE_PYTHON in the .Renviron.
