+++ R/ee_print.R
@@ -11,6 +11,8 @@
 #' Relevant just for \code{ee$ImageCollection} objects.
 #' @param f_index Numeric. Index of the \code{ee$FeatureCollection} to fetch.
 #' Relevant just for \code{ee$FeatureCollection} objects.
+#' @param time_end Logical. If TRUE, the system:time_end property in ee$Image
+#' is also returned. See \code{rgee::ee_get_date_img} for details.
 #' @param img_band Character. Band name of the \code{ee$Image} to fetch.
 #' Relevant just for \code{ee$ImageCollection} and \code{ee$Image} objects.
 #' @param compression_ratio Numeric. Measurement of the relative reduction
@@ -31,8 +33,6 @@
 #' @examples
 #' \dontrun{
 #' library(rgee)
-#'
-#' ee_reattach()
 #' ee_Initialize()
 #'
 #' # Geometry
@@ -93,13 +93,14 @@ ee_print.ee.geometry.Geometry <- function(eeobject,
     geom_type <- toupper(geometry$type)
     geom_info <- eeobject$projection()$getInfo()
     geom_epsg <- as.numeric(gsub("EPSG:", "", geom_info$crs))
+    geom_geodesic <- ee_utils_py_to_r(ee$Geometry$geodesic(eeobject)$getInfo())
     geom_proj4string <- st_crs(geom_epsg)$proj4string
     geom_geotransform <- paste0(geom_info$transform, collapse = " ")
-
     ee_metadata <- list(
       name = "Geometry",
       geom_type = geom_type,
       geom_epsg = geom_epsg,
+      geom_geodesic = geom_geodesic,
       geom_geotransform = geom_geotransform,
       geom_proj4string = geom_proj4string
     )
@@ -117,6 +118,7 @@ ee_print.ee.geometry.Geometry <- function(eeobject,
     cat("\n - EPSG (SRID)                :", ee_metadata$geom_epsg)
     cat("\n - proj4string                :", ee_metadata$geom_proj4string)
     cat("\n - Geotransform               :", ee_metadata$geom_geotransform)
+    cat("\n - Geodesic                   :", ee_metadata$geom_geodesic)
     cat("\n - WKT                        :", ee_metadata$geom_type)
     cat("\n", rule())
   }
@@ -148,6 +150,7 @@ ee_print.ee.feature.Feature <- function(eeobject,
     # 3. Geometry metadata
     geom_type <- toupper(feature_info$geometry$type)
     geom_info <- eeobject$geometry()$projection()$getInfo()
+    geom_geodesic <- ee_utils_py_to_r(eeobject$geometry()$geodesic()$getInfo())
     geom_epsg <- as.numeric(gsub("EPSG:", "", geom_info$crs))
     geom_proj4string <- st_crs(geom_epsg)$proj4string
     geom_geotransform <- paste0(geom_info$transform, collapse = " ")
@@ -158,6 +161,7 @@ ee_print.ee.feature.Feature <- function(eeobject,
       f_properties_length = f_properties_length,
       geom_type = geom_type,
       geom_epsg = geom_epsg,
+      geom_geodesic = geom_geodesic,
       geom_geotransform = geom_geotransform,
       geom_proj4string = geom_proj4string
     )
@@ -177,6 +181,7 @@ ee_print.ee.feature.Feature <- function(eeobject,
     cat("\n - EPSG (SRID)                :", ee_metadata$geom_epsg)
     cat("\n - proj4string                :", ee_metadata$geom_proj4string)
     cat("\n - Geotransform               :", ee_metadata$geom_geotransform)
+    cat("\n - Geodesic                   :", ee_metadata$geom_geodesic)
[…]
