+++ R/ee_imagecollection.R
@@ -37,8 +37,6 @@
 #' \dontrun{
 #' library(rgee)
 #' library(raster)
-#'
-#' ee_reattach()
 #' ee_Initialize(drive = TRUE, gcs = TRUE)
 #'
 #' # USDA example
