### 2020-06-10 296aae3
working in rgee 0.6.0
