### 2020-06-13 8ae9413
rgee 0.6.0 release
