+++ tests/testthat/test-map.R
@@ -194,3 +194,4 @@ test_that("messages 01", {
 })
 
 
+
