+++ R/ee_Map.R
@@ -1,548 +1,0 @@
-#' Module to display Earth Engine (EE) spatial objects
-#'
-#' Create interactive visualizations of spatial EE objects
-#' (ee$Geometry, ee$Image, ee$Feature, and ee$FeatureCollection)
-#' using \code{mapview}.
-#' @format An object of class environment with the
-#' following functions:
-#' \itemize{
-#'   \item  \strong{addLayer(eeObject, visParams, name = NULL, shown = TRUE,
-#'   opacity = 1)}: Adds a given EE object to the map as a layer. \cr
-#'   \itemize{
-#'     \item \strong{eeObject:} The object to add to mapview.\cr
-#'     \item \strong{visParams:} List of parameters for visualization.
-#'     See details.\cr
-#'     \item \strong{name:} The name of the layer.\cr
-#'     \item \strong{shown:} A flag indicating whether the
-#'     layer should be on by default. \cr
-#'     \item \strong{opacity:} The layer's opacity represented as a number
-#'      between 0 and 1. Defaults to 1. \cr
-#'   }
-#'   \item \strong{setCenter(lon = 0, lat = 0, zoom = NULL)}: Centers the map
-#'   view at the given coordinates with the given zoom level. If no zoom level
-#'   is provided, it uses 1 by default.
-#'   \itemize{
-#'     \item \strong{lon:} The longitude of the center, in degrees.\cr
-#'     \item \strong{lat:} The latitude of the center, in degrees.\cr
-#'     \item \strong{zoom:} The zoom level, from 1 to 24.
-#'   }
-#'   \item \strong{setZoom(zoom = NULL)}: Sets the zoom level of the map.
-#'   \itemize{
-#'     \item \strong{zoom:} The zoom level, from 1 to 24.
-#'   }
-#'   \item \strong{centerObject(eeObject, zoom = NULL,
-#'    maxError = ee$ErrorMargin(1))}: Centers the
-#'   map view on a given object. If no zoom level is provided, it will
-#'   be predicted according to the bounds of the Earth Engine object specified.
-#'   \itemize{
-#'     \item \strong{eeObject:} EE object.\cr
-#'     \item \strong{zoom:} The zoom level, from 1 to 24.
-#'     \item \strong{maxError:} 	Max error when input
-#'     image must be reprojected to an explicitly
-#'     requested result projection or geodesic state.
-#'   }
-#' }
-#'
-#' @details
-#' `Map` use the Earth Engine method
-#' \href{https://developers.google.com/earth-engine/api_docs#ee.data.getmapid}{
-#' getMapId} to fetch and return an ID dictionary being used to create
-#' layers in a \code{mapview} object. Users can specify visualization
-#' parameters to Map\$addLayer by using the visParams argument. Each Earth
-#' Engine spatial object has a specific format. For
-#' \code{ee$Image}, the
-#' \href{https://developers.google.com/earth-engine/image_visualization}{
-#' parameters} available are:
-#'
-#' \tabular{lll}{
-#' \strong{Parameter}\tab \strong{Description}  \tab \strong{Type}\cr
-#' \strong{bands}    \tab  Comma-delimited list of three band names to be
-#' mapped to RGB     \tab  list \cr
-#' \strong{min}      \tab  Value(s) to map to 0 \tab  number or list of three
-#' numbers, one for each band \cr
[…]
