+++ .github/workflows/updated.yaml
@@ -45,26 +45,6 @@ jobs:
         env:
           RGEE_SECRET_PASSPHRASE: ${{ secrets.RGEE_SECRET_PASSPHRASE }}
 
-      - name: INSTALL SYSTEM DEPENDENCIES [macOS]
-        if: runner.os == 'macOS'
-        env:
-          RHUB_PLATFORM: osx-x86_64-clang
-        run: |
-          brew install udunits
-          brew install gdal
-          brew install v8
-          brew install freetype
-          cat <<EOT >> .Renviron
-          PKG_CONFIG_PATH=/usr/local/lib/pkgconfig/
-          PROJ_LIB=/usr/local/opt/proj/share/proj/
-          EOT
-          cat <<EOT >> .Rprofile
-          config_args <- c("sf" = "--with-proj-lib=/usr/local/lib/", "rgdal" = "--with-proj-lib=/usr/local/lib/ --with-proj-include=/usr/local/include/")
-          r <- getOption("repos")
-          r["CRAN"] <- "https://cran.rstudio.com"
-          options(configure.args = config_args, repos = r)
-          EOT
-
       - name: INSTALL PYTHON DEPENDENCIES
         run: |
           pip install --upgrade pip
@@ -91,11 +71,6 @@ jobs:
       - name: PREBUILD TESTING
         run: |
           earthengine -h
-
-      - name: INSTALL rgdal WITH CONFIGURE
-        run: |
-          install.packages("rgdal", type = "source", configure.args = "--with-proj-lib=/usr/local/lib/ --with-proj-include=/usr/local/include/")
-        shell: Rscript {0}
 
       - name: UPGRADE RGEE
         run: Rscript ./.github/updated.R
