+++ .github/workflows/R-CMD-check.yaml
@@ -66,37 +66,6 @@ jobs:
           close(fileConn)
         shell: Rscript {0}
 
-      - name: INSTALL SYSTEM DEPENDENCIES [LINUX]
-        if: runner.os == 'Linux'
-        env:
-          RHUB_PLATFORM: ${{ matrix.config.platform }}
-        run: |
-          sudo apt update
-          sudo add-apt-repository -y ppa:ubuntugis/ubuntugis-unstable
-          sudo apt install libudunits2-dev libgdal-dev libgeos-dev libproj-dev \
-          libv8-3.14-dev libjq-dev libnetcdf-dev libjson-c-dev libfontconfig1-dev \
-          libprotobuf-dev protobuf-compiler
-
-      - name: INSTALL SYSTEM DEPENDENCIES [macOS]
-        if: runner.os == 'macOS'
-        env:
-          RHUB_PLATFORM: ${{ matrix.config.platform }}
-        run: |
-          brew install udunits
-          brew install freetype
-          brew install gdal
-          brew install protobuf
-          cat <<EOT >> .Renviron
-          PKG_CONFIG_PATH=/usr/local/lib/pkgconfig/
-          PROJ_LIB=/usr/local/opt/proj/share/proj/
-          EOT
-          cat <<EOT >> .Rprofile
-          config_args <- c("sf" = "--with-proj-lib=/usr/local/lib/", "rgdal" = "--with-proj-lib=/usr/local/lib/ --with-proj-include=/usr/local/include/")
-          r <- getOption("repos")
-          r["CRAN"] <- "https://cran.rstudio.com"
-          options(configure.args = config_args, repos = r)
-          EOT
-
       - name: INSTALL PYTHON DEPENDENCIES
         run: |
           pip install --upgrade pip
@@ -113,12 +82,6 @@ jobs:
       - name: PREBUILD TESTING
         run: |
           earthengine -h
-
-      - name: INSTALL sfrgdal WITH CONFIGURE [macOS]
-        if: runner.os == 'macOS'
-        run: |
-          install.packages("rgdal", type = "source", configure.args = "--with-proj-lib=/usr/local/lib/ --with-proj-include=/usr/local/include/")
-        shell: Rscript {0}
 
       - name: INSTALL dependencies and rcmdcheck
         run: |
