+++ pkgdown.yml
@@ -5,7 +5,7 @@ articles:
   BestPractices: BestPractices.html
   considerations: considerations.html
   setup: setup.html
-last_built: 2020-06-13T05:15Z
+last_built: 2020-06-13T05:28Z
 urls:
   reference: http://r-spatial.github.io/rgee/reference
   article: http://r-spatial.github.io/rgee/articles
