### 2020-06-13 261514b
rgee 0.6.0 release
