+++ .github/workflows/updated.yaml
@@ -1,10 +1,10 @@
 on:
   push:
     branches:
-      - dev
+      - master
   pull_request:
     branches:
-      - dev
+      - master
   schedule:
     - cron: "0 1 * * *"
 
