+++ .github/workflows/test-coverage.yaml
@@ -1,10 +1,10 @@
 on:
   push:
     branches:
-      - dev
+      - master
   pull_request:
     branches:
-      - dev
+      - master
 
 name: test-coverage
 
@@ -78,7 +78,7 @@ jobs:
 
       - name: INSTALL R TEST PACKAGES
         run: |
-          install.packages(c("remotes", "Rcpp")
+          install.packages(c("remotes", "Rcpp"))
           remotes::install_deps(dependencies = TRUE)
           remotes::install_cran("covr", force = TRUE)
           remotes::install_cran("rcmdcheck")
