+++ .github/workflows/pkgdown.yaml
@@ -30,7 +30,6 @@ jobs:
       - name: Install dependencies
         run: |
           install.packages("remotes")
-          remotes::install_deps(dependencies = TRUE)
           remotes::install_cran("pkgdown")
         shell: Rscript {0}
 
