+++ .github/workflows/pkgdown.yaml
@@ -17,7 +17,7 @@ jobs:
         run: |
           install.packages("Rcpp")
           install.packages('remotes')
-          saveRDS(remotes::dev_package_deps(dependencies = TRUE), ".github/depends.Rds", version = 2)
+          saveRDS(remotes::dev_package_deps(dependencies = "Imports"), ".github/depends.Rds", version = 2)
         shell: Rscript {0}
 
       - name: Cache R packages
@@ -30,6 +30,7 @@ jobs:
       - name: Install dependencies
         run: |
           install.packages("remotes")
+          remotes::install_deps(dependencies = "Imports")
           remotes::install_cran("pkgdown")
         shell: Rscript {0}
 
