+++ .github/workflows/R-CMD-check.yaml
@@ -52,9 +52,10 @@ jobs:
       - name: FIND ALL R PACKAGES DEPENDECIES OF RGEE
         run: |
           install.packages('remotes')
+          install.packages('knitr')
           install.packages('pillar') # winbug
           install.packages('vctrs')
-          saveRDS(remotes::dev_package_deps(dependencies = TRUE), ".github/depends.Rds", version = 2)
+          saveRDS(remotes::dev_package_deps(dependencies = "Imports"), ".github/depends.Rds", version = 2)
         shell: Rscript {0}
 
       - name: RETICULATE_PYTHON SETTING
@@ -85,7 +86,7 @@ jobs:
 
       - name: INSTALL dependencies and rcmdcheck
         run: |
-          remotes::install_deps(dependencies = TRUE)
+          remotes::install_deps(dependencies = "Imports")
           remotes::install_cran("rcmdcheck")
         shell: Rscript {0}
 
