### 2020-06-14 2cbe73f
rgee 0.6.1 release
