+++ R/ee_version.R
@@ -5,5 +5,5 @@
 #' @return Character. Earth Engine Python API version used to build rgee.
 #' @export
 ee_version <- function() {
- '0.1.225'
+ '0.1.226'
 }
