### 2020-06-19 020361e
ee_check was completely renovated
- new message display
- Fix a bug in test GCS credentials.
- ee_check.py is no longer necessary.
