+++ R/ee_check.R
@@ -1,11 +1,13 @@
 #' Interface to check Python and non-R dependencies
 #'
 #' R functions for checking Google credentials (Google Earth Engine,
-#' Google Drive and Google Cloud Storage) available in their system and the
-#' Python environment and the Third-Party Python Packages used by rgee.
+#' Google Drive and Google Cloud Storage), Python environment and the
+#' Third-Party Python Packages used by rgee.
 #'
 #' @name ee_check-tools
 #'
+#' @param user Character. User to check credentials. If it is not defined,
+#' ee_check will skip the check of credentials.
 #' @param quiet Logical. Suppress info message
 #'
 #' @importFrom reticulate py_available py_module_available py_discover_config
@@ -20,13 +22,15 @@
 #' ee_check_python()
 #' ee_check_python_packages()
 #' ee_check_credentials()
-#' ee_check() # put it all together
+#' ee_check() # put them all together
 #' }
 #' @export
-ee_check <- function() {
-  ee_check_python()
-  ee_check_python_packages()
-  ee_check_credentials()
+ee_check <- function(user = NULL, quiet = FALSE) {
+  ee_check_python(quiet = quiet)
+  ee_check_python_packages(quiet = quiet)
+  if (!is.null(user)) {
+    ee_check_credentials(quiet = quiet)
+  }
 }
 
 #' @rdname ee_check-tools
@@ -39,8 +43,10 @@ ee_check_python <- function(quiet = FALSE) {
     if (!quiet) {
       cat_line(
         blue(symbol$circle_filled),
-        blue("  Python version found: "),
-        green(
+        blue("  Python version\n"),
+        green(symbol$tick, "[Ok]"),
+        blue(
+          "",
           py_discover_config()$python,
           sprintf("v%s", py_version)
         )
@@ -48,7 +54,7 @@ ee_check_python <- function(quiet = FALSE) {
     }
   } else {
     stop(
-      "Unable to find a Python version, you will need to fix it before run ",
+      "Unable to find a Python version, you will need to fix before run ",
       "rgee::ee_Initialize(). For more details run reticulate::py_available()"
     )
   }
@@ -60,101 +66,86 @@ ee_check_python <- function(quiet = FALSE) {
 #' @family ee_check functions
 #' @export
 ee_check_python_packages <- function(quiet = FALSE) {
-  oauth_func_path <- system.file("python/ee_check.py",
-    package = "rgee"
-  )
-  ee_check_utils_exist <- ee_source_python(oauth_func_path)
+  # Necessary packages
+  if (!py_module_available("numpy") & quiet == FALSE) {
+    ee_wrong_message("numpy")
+    return(invisible(FALSE))
+  }
+  if (!py_module_available("ee") & quiet == FALSE) {
+    ee_wrong_message("earthengine-api")
+    return(invisible(FALSE))
+  }
   if (!quiet) {
     cat_line(
-      "\n",
-      blue(
-        symbol$circle_filled,
-        " Python Third-Party Libraries used in rgee: \n"
-      )
+      blue(symbol$circle_filled),
+      blue("  Python packages:\n"),
+      green(symbol$tick, "[Ok]"),
+      blue(" numpy\n"),
+      green(symbol$tick, "[Ok]"),
+      blue(" earthengine-api")
     )
   }
-  # ee
-  version_ee <- ee_utils_py_to_r(ee_check_utils_exist$ee_check_py_ee())
-  ee_cond <- is.character(version_ee)
-  if (ee_cond) {
-    if (version_ee == ee_version()) {
-      if (!quiet) {
-        cat_line(
-          green(symbol$tick, "[Ok]"),
-          blue(symbol$check, "Python Earth Engine API version "),
[…]
