### 2020-06-19 1d74d28
ee_Initialize bug fixed
Fix a bug when switching between users with a different number of uploaded credentials
