+++ tests/testthat/test-check.R
@@ -47,13 +47,11 @@ init_rgee()
 # -------------------------------------------------------------------------
 test_that("simple ee_check ",{
   expect_true(ee_check_python())
-  expect_null(ee_check_python_packages())
+  expect_true(ee_check_python_packages())
   expect_true(ee_check_credentials())
 })
 
 test_that("ee_check ",{
-  expect_true(ee_check())
+  expect_null(ee_check())
 })
-
-
 
