+++ DESCRIPTION
@@ -43,16 +43,20 @@ Authors@R:
              family = "Tang",
              role = c("ctb"),
              email = "terrytangyuan@gmail.com"),
-      person(given = "Gabriel",
-             family = "Carrasco",
+      person(given = "Samapriya",
+             family = "Roy",
              role = c("ctb"),
-             email = "gabriel.carrasco@upch.pe",
-             comment = c(ORCID = "0000-0002-6945-0419")),
+             email = "samapriya.roy@gmail.com"),
       person(given = "MariaElena",
              family = "Adauto",
              role = c("ctb"),
              email = "2a.mariaelena@gmail.com",
              comment = c(ORCID = "0000-0002-2154-2429")),
+      person(given = "Gabriel",
+             family = "Carrasco",
+             role = c("ctb"),
+             email = "gabriel.carrasco@upch.pe",
+             comment = c(ORCID = "0000-0002-6945-0419")),
       person(given = "Jeffrey",
              family = "Hollister",
              role = c("rev"),
