### 2020-06-21 81bde4f
rgee 0.6.2 release
