### 2020-06-22 454eb81
now mapview and EarthEnginemap can interact each other!
