### 2020-06-25 r-spatial/rgee#68 @jhollist
Review for rgee JOSS Submission
@csaybar see below for my review for your submission at https://github.com/openjournals/joss-reviews/issues/2272

# General Comments

This is a very nicely written and conceived package, paper, and repository!  I had not used Google Earth Engine and could see myself giving it a try because of `rgee`.  So my first comment is "Nice Job!"

Second, the paper is also very well written.  No major comments on the text of the paper other than it feels a little on the long side for a JOSS paper, but that being said I don't think there is much I would suggest removing.  If JOSS prefers a bit more brevity, maybe you could reduce some of the examples and point to those examples that you had in the README.

Third,  the documentation in the repository is also very well done and again no major comments from me.

I do have a few things that came up while I was working through the code that I think should be addressed.  They all map on to the items in the checklist that I haven't marked off yet.

# Checklist comments

- *Contribution and authorship:*  Authors on paper look good for the most part.  Did check commits on repo and don't see Qiusheng Wu on the repo, and dont see Gabriel Carrasco-Escobar who made commits.  Not questioning your author choices, just curious about the differences.
- *Installation:* For the most part the installation went fairly smoothly.  You have done a very nice job dealing with the install of the needed python dependencies.  I did have a little issue getting the first `ee_Initialize()` working as it requires an Earth Engine account.   I am new to earth engine and as such did not already have an account.  The error messages were useful in tracing this down.  Couple of suggestions.
  - Indicate somewhere, maybe message early on with ee_Initialize(), that a EE account is needed.  Perhaps ask if user has an EE account before the auth?
  - I did not have the root folder set up so got the error about this.  Maybe include the path in quotes inside of  `ee$data$createAssetHome(users/PUT_YOUR_NAME_HERE)` so example in error reads `ee$data$createAssetHome("users/PUT_YOUR_NAME_HERE")`
  - If a user folder with the name you choose already exists, then you get the following error from python:

>Error in py_call_impl(callable, dots$args, dots$keywords) :
>EEException: User home folder already exists: users/jhollist.

- Perhaps check for existence first as I could see many users wanting to use their existing user folder
- After navigating these small issues, was able to get everything up and running.

- *References:*  Two of your references have DOIs and should be included in the citation:
  - Gorelick: https://doi.org/10.1016/j.rse.2017.06.031
  - Wickham: https://doi.org/10.1002/wics.147

# Miscellaneous comments

- This is me being nit-picky, but in the Extract Precipitation Values example you use `library(tidyverse)`.  I am of the opinion it is better to load only those packages that are being used for a given analysis.  So in this case, I'd replace `library(tidyverse)` with `library(tidyr)` and `library(ggplot2)`.

- Version: Latest release on GitHub is v0.2.0, DESCRIPTION has version at 0.6.2, and submitted version with JOSS is v0.5.0.  Just make sure these all are synced up.
