### 2020-06-25 #2272 @jhollist
@csaybar Really nice job!  I have posted a new issue on your repo (https://github.com/r-spatial/rgee/issues/68) with some more detailed review comments.  Any questions, let me know.
