+++ paper/paper.bib
@@ -5,6 +5,7 @@
   volume={202},
   pages={18--27},
   year={2017},
+  doi={10.1016/j.rse.2017.06.031},
   publisher={Elsevier}
 }
 
@@ -24,20 +25,9 @@
   number={2},
   pages={180--185},
   year={2011},
+  doi={10.1016/j.rse.2017.06.031},
   publisher={Wiley Online Library}
 }
-
-
-@article{abatzoglou2018terraclimate,
-  title={TerraClimate, a high-resolution global dataset of monthly climate and climatic water balance from 1958--2015},
-  author={Abatzoglou, John T and Dobrowski, Solomon Z and Parks, Sean A and Hegewisch, Katherine C},
-  journal={Scientific data},
-  volume={5},
-  pages={170191},
-  year={2018},
-  publisher={Nature Publishing Group}
-}
-
 
 @Manual{reticulate,
   author = {JJ Allaire, Kevin Ushey, Yuan Tang and Dirk Eddelbuettel},
