+++ man/Map.Rd
@@ -139,6 +139,7 @@ m4 <- Map$addLayer(
 m4
 
 # Case 5: mapview + EarthEnginemap
+library(sf)
 nc <- st_read(system.file("shp/arequipa.shp", package="rgee"))
 mapview(nc) + m2
 m2 + mapview(nc)
