+++ NAMESPACE
@@ -75,12 +75,12 @@ export(eedate_to_rdate)
 export(gcs_to_ee_image)
 export(gcs_to_ee_table)
 export(local_to_gcs)
-export(print.ee.computedobject.ComputedObject)
 export(raster_as_ee)
 export(rdate_to_eedate)
 export(sf_as_ee)
 export(stars_as_ee)
 exportClasses(EarthEngineMap)
+exportClasses(mapview)
 importFrom(cli,cat_line)
 importFrom(cli,rule)
 importFrom(cli,symbol)
