+++ paper/paper.bib
@@ -25,7 +25,7 @@
   number={2},
   pages={180--185},
   year={2011},
-  doi={10.1016/j.rse.2017.06.031},
+  doi={10.1002/wics.147},
   publisher={Wiley Online Library}
 }
 
