### 2020-06-26 r-spatial/rgee#68 @csaybar
Hi @jhollist thank you so much for taking your time in review rgee.  I implemented the changes in #69 and recompiled the paper (see https://github.com/r-spatial/rgee/blob/JOSS/paper/paper.md). I commented on all your suggestions. Please let me know if there are any issues remaining. I will keep #68 open until all the review is completed.

Have a nice day!

> Contribution and authorship: Authors on paper look good for the most part. Did check commits on repo and don't see Qiusheng Wu on the repo, and dont see Gabriel Carrasco-Escobar who made commits. Not questioning your author choices, just curious about the differences.

All [rgee examples](https://csaybar.github.io/rgee-examples/) are an identical copy of Prof. Qiusheng Wu repositories (See [earthengine-py-notebooks](https://github.com/giswqs/earthengine-py-notebooks) and [qgis-earthengine-examples](https://github.com/giswqs/qgis-earthengine-examples), besides many rgee functions are based/inspired by [geemap](https://github.com/giswqs/geemap) (e.g. Map$adLayer). Prof. Qiusheng Wu also contributed a lot in the generation of the document. Unfortunately, due to the fact that we worked on it through google docs there is no record of his work. With respect to @barja8 and @LBautistaB13, they generated most of the tests we currently use on rgee. Specifically, @LBautistaB13 worked several days on understanding and after that teaching me how Github actions works :relaxed:, @barja8 and @ryali93 make that everything works well in Windows. However, they were not very comfortable with work with branches and PR (they wanted to avoid responsibility if something broke) so there is no sign of all their big effort in the repository either. Gabriel Carrasco-Escobar gave a fast check of function documentation and made some commits, however, because he didn't contribute the design nor in the code, we considered adding him to the package only as a contributor.

> Indicate somewhere, maybe message early on with ee_Initialize(), that a EE account is needed. Perhaps ask if user has an EE account before the auth?

Good point, thank you. Now all new rgee users will see this message.

> I did not have the root folder set up so got the error about this. Maybe include the path in quotes inside of ee$data$createAssetHome(users/PUT_YOUR_NAME_HERE) so example in error reads ee$data$createAssetHome("users/PUT_YOUR_NAME_HERE")

Done.

> If a user folder with the name you choose already exists, then you get the following error from python:

Thank you for pointing this out. It is a scenario that we hadn't considered. We implement `ee_createAssetHome` to help users to define their Earth Engine Assets home root folder without leaving `ee_Initialize()`

> References: Two of your references have DOIs and should be included in the citation:
> Gorelick: https://doi.org/10.1016/j.rse.2017.06.031
> Wickham: https://doi.org/10.1002/wics.147

Done.

> This is me being nit-picky, but in the Extract Precipitation Values example you use library(tidyverse). I am of the opinion it is better to load only those packages that are being used for a given analysis. So in this case, I'd replace library(tidyverse) with library(tidyr) and library(ggplot2).

Thank you we made the change.

> Version: Latest release on GitHub is v0.2.0, DESCRIPTION has version at 0.6.2, and submitted version with JOSS is v0.5.0. Just make sure these all are synced up.

Good point, this was fixed
