### 2020-06-26 #2272 @jhollist
@hugoledoux and @csaybar  I have completed the checklist and the edits (https://github.com/r-spatial/rgee/pull/69) made in response to my review (https://github.com/r-spatial/rgee/issues/68) all look great.  Again, really nice job on this package.
