### 2020-06-26 r-spatial/rgee#68 @jhollist
@csaybar Edits look good.  I have finished up the checklist at https://github.com/openjournals/joss-reviews/issues/2272  and as far as my review is concerned, you are all set!  I look forward to using `rgee` for some of my projects.
