### 2020-06-26 #2272 @csaybar
Thank @jhollist for your comments and suggestions :). They were addressed here r-spatial/rgee#69
