+++ R/sf_as_ee.R
@@ -10,7 +10,6 @@
 #' will be monitored.
 #' @param assetId Character. Destination asset ID for the uploaded file. Ignore
 #' if \code{via} argument is "getInfo".
-#' @param check_ring_dir Logical. See \code{st_read} for details.
 #' @param proj Integer or character. Coordinate Reference System (CRS) for the
 #' EE object, defaults to "EPSG:4326" (x=longitude, y=latitude).
 #' @param geodesic Logical. Ignored if \code{x} is not a Polygon or LineString.
@@ -26,6 +25,9 @@
 #' If unspecified, defaults to TRUE.
 #' @param bucket Character. Name of the bucket (GCS) to save intermediate files
 #' (ignore if \code{via} is not defined as "gcs_to_asset").
+#' @param command_line_tool_path Character. Path to the Earth Engine command line
+#' tool. If NULL, rgee assumes is saved in the same path that your Python
+#' environment (ignore if \code{via} is not defined as "gcs_to_asset").
 #' @param overwrite A boolean argument which indicates indicating
 #' whether "filename" should be overwritten. By default TRUE.
 #' @param quiet Logical. Suppress info message.
@@ -55,12 +57,6 @@
 #' See \href{https://developers.google.com/earth-engine/importing}{Importing
 #' table data} documentation for more details.
 #'
-#' Earth Engine is strict on polygon ring directions (outer ring
-#' counter-clockwise, and the inner one clockwise). If `check_ring_dir` is TRUE,
-#' it checks every ring, and revert them if necessary, to counter clockwise for
-#' outer, and clockwise for inner (hole) ones. By default, this is FALSE because
-#' it is an expensive operation.
-#'
 #' @examples
 #' \dontrun{
 #' library(rgee)
@@ -80,7 +76,10 @@
 #'                    ncol = 2,
 #'                    byrow = TRUE) %>%
 #'   list() %>%
-#'   st_polygon()
+#'   st_polygon() %>%
+#'   st_sfc(crs="EPSG:4326") %>%
+#'   st_sf()
+#'
 #' holePoly <- sf_as_ee(x = toy_poly, evenOdd = FALSE)
 #'
 #' # Create an even-odd version of the polygon.
@@ -99,10 +98,11 @@
 #'   x = toy_poly,
 #'   overwrite = TRUE,
 #'   assetId = assetId,
-#'   via = 'getInfo_to_asset')
+#'   via = "getInfo_to_asset")
 #'
 #' # 3. Upload large geometries to EE asset
-#' ee_Initialize(gcs = TRUE)
+#' ee_Initialize("data.colec.fbf", gcs = TRUE)
+#'
 #' assetId <- sprintf("%s/%s", ee_get_assethome(), 'toy_poly_gcs')
 #' eex <- sf_as_ee(
 #'   x = toy_poly,
@@ -112,38 +112,38 @@
 #'   monitoring = FALSE,
 #'   via = 'gcs_to_asset'
 #' )
+#' ee_monitoring()
 #' }
 #' @export
 sf_as_ee <- function(x,
                      via = 'getInfo',
                      assetId = NULL,
                      bucket = NULL,
+                     command_line_tool_path = NULL,
                      overwrite = TRUE,
                      monitoring = TRUE,
-                     check_ring_dir = FALSE,
                      evenOdd = TRUE,
                      proj = 4326,
                      geodesic = NULL,
                      quiet = FALSE,
                      ...) {
 
-  # Read geometry
-  eex <- ee_st_read(
-    x = x,
[…]
