### 2020-06-30 r-spatial/rgee#70 @appelmar
JOSS Review
Great work @csaybar! Here is my review of the JOSS submission (openjournals/joss-reviews#2272).

# General Comments

The rgee package is a welcome addition to the R spatial ecosystem. It allows to combine the processing of massive Earth observation datasets in Google Earth Engine (GEE) with R-based analysis workflows (e.g., to use R for pre/post-processing, visualization, and mapping). The amount of examples available online (https://csaybar.github.io/rgee-examples/) is impressive and very helpful for users to familiarize with GEE / rgee concepts.

All in all, the package and the paper are well written and the documentation is very comprehensive. The package clearly makes it *very easy* to use GEE from R. Below, please find some minor(!) comments for the paper and the software separately.

# Paper

The paper highlights most important design decisions of the package and includes helpful code examples. I only have two minor comments:

- The source code comment in the time series extraction example _"Image or ImageCollection (mean composite)"_ is not so clear, I'd suggest to describe what you actually do in the lines below (select dataset, reproject, filter by date and variable).

- The paper mentions that rgee is available from CRAN but it is not (yet).

# Software

- Installation (including a Miniconda Python environment) went smooth on my platform (Ubuntu 18.04, R 4.0.1, and rgee package version 0.6.2, commit 454eb81102802813516ce7fb21986737e37e5581).

- For the `region` argument in  `ee_as_stars()` and `ee_as_raster()`, the documentation states that "if not specified, image bounds will be taken". However, in my tests, I get _"Error in ee_image_local(image = image, region = region, dsn = dsn, via = via,  : argument "region" is missing, with no default"_ when calling these functions without region argument.

- I've found that calling `ee_Initialize(drive = TRUE, gcs = TRUE)` sometimes constantly asks for a selection of the mail address (though there is only one listed). If selecting 1, it asks again whereas if selecting 0, the package opens the browser for authorization, and after allowing everything, it asks again. Am I doing something wrong? I don't know any Google API details but if possible, I think it would be nice to make the Google drive / cloud storage authorization as simple as when calling `ee_Initialize()` for GEE only (which works well).

- It is clear that function examples in the documentation are marked as `dontrun`, because they need a GEE account. However, I would have enjoyed to run R CMD check with argument `--run-dontrun` with my own GEE acount, and also run the testthat tests. I was not successful because it seems that the GEE user is hard-coded in the tests and some examples (with Google drive or cloud storage) do not work in a non-interactive session. Do you see any way to make it easier for others to run all tests / examples automatically?

- `R CMD check --as-cran` reports some documentation issues of `print.ee.computedobject.ComputedObject()`: _"Found the following apparent S3 methods exported but not registered: print.ee.computedobject.ComputedObject"_

- Nice work on the `ee_help()` function! It is very helpful to have the GEE documentation of functions in the R help viewer.
