### 2020-06-30 #2272 @appelmar
Great work @csaybar. Please find my detailed review with some minor comments in a separate issue in your repo (see  r-spatial/rgee#70). Let me know if you have questions.
