### 2020-06-30 #2272 @appelmar
@hugoledoux I can't fill out the checklist in this issue, did I miss something?
