### 2020-07-03 r-spatial/rgee#72 @gena
JOSS Review
This is a very nice R package @csaybar and will be for sure useful among R users for geospatial processing combining local analysis with the Cloud-based Google Earth Engine (GEE) processing.

I've followed instructions provided in the ```README.md``` - everything went smoothly and installed without issues on Windows 10.

* The second [example](https://github.com/r-spatial/rgee#2-extract-precipitation-values) mentions ```shape/nc.shp``` file, but there is no link to where the file can be downloaded - please share it somewhere or add to the repository if it's tiny. This would help users to follow instructions and reproduce your examples 100%. The same is also true for the shp/arequipa.shp in the [second example](https://github.com/r-spatial/rgee#3-create-an-ndvi-animation-js-version).

The documentation on the GitHub is very informative and useful and helps to get started fast with the package.
