### 2020-07-03 #2272 @gena
I've added my review https://github.com/r-spatial/rgee/issues/72, it's short, not a lot to add to other reviews.
