+++ tests/testthat/test-thumbnail.R
@@ -1,54 +1,6 @@
 context("rgee: ee_as_thumbnail test")
-
-# Pre-checking ------------------------------------------------------
-# Google credentials were loaded in the system?
-skip_if_no_credentials <- function(user) {
-  ee_path <- path.expand(sprintf("~/.config/earthengine/%s", user))
-  credentials <- list.files(
-    path = ee_path,
-    pattern = "@gmail.com|credentials|GCS_AUTH_FILE.json"
-  )
-  if (length(credentials) != 3) {
-    skip("All google credentials were not found")
-  }
-}
-
-# Neccesary Python packages were loaded?
-skip_if_no_pypkg <- function() {
-  have_ee <- reticulate::py_module_available("ee")
-  have_numpy <- reticulate::py_module_available("numpy")
-  if (isFALSE(have_ee)) {
-    skip("ee not available for testing")
-  }
-  if (isFALSE(have_numpy)) {
-    skip("numpy not available for testing")
-  }
-}
-
-# Init Earth Engine just if it is necessary
-init_rgee <- function() {
-  tryCatch(
-    expr = ee$Image()$getInfo(),
-    error = function(e) {
-      ee_Initialize(
-        email = 'data.colec.fbf@gmail.com',
-        drive = TRUE,
-        gcs = TRUE
-      )
-    }
-  )
-}
-
-user <- "data.colec.fbf"
-skip_if_no_credentials(user)
 skip_if_no_pypkg()
-init_rgee()
 # -------------------------------------------------------------------------
-
-
-library(raster)
-library(rgee)
-library(sf)
 
 ### 1. Data
 dem_palette <- c(
@@ -63,6 +15,7 @@ sheds <- ee$FeatureCollection("USGS/WBD/2017/HUC06")$
   num <- ee$Number$parse(feature$get("areasqkm"))
   return(feature$set("areasqkm", num))
 })
+
 image <- ee$Image("CGIAR/SRTM90_V4")
 region <- nc$geometry[[1]] %>%
   st_bbox() %>%
@@ -107,7 +60,6 @@ test_that("ee_as_thumbnail palette, min-max", {
   expect_equal(max(arequipa_dem[[1]]), 5000, tolerance = 1)
 })
 
-
 # RGB band -----------------------------------------------------------
 test_that("ee_as_thumbnail region", {
   # PNG images
@@ -137,12 +89,10 @@ test_that("ee_as_thumbnail error 02", {
   expect_error(ee_as_thumbnail(image, "ee$Geometry"))
 })
 
-
 test_that("ee_as_thumbnail error 03", {
   # PNG images
   expect_error(ee_as_thumbnail(image, region$centroid()$buffer(100)))
 })
-
 
 # large image -----------------------------------------------------------
 test_that("ee_as_thumbnail large image", {
@@ -155,5 +105,3 @@ test_that("ee_as_thumbnail large image", {
   )
   expect_s3_class(arequipa_dem, "stars")
 })
-
-

[…]
