### 2020-07-05 r-spatial/rgee#70 @csaybar
Hi @appelmar  thank you so much for taking your time in review rgee. I implemented the changes in #69 (See [NEWS.md](https://github.com/r-spatial/rgee/blob/JOSS/NEWS.md)) and recompiled the paper (see https://github.com/r-spatial/rgee/blob/JOSS/paper/paper.md). I commented on all your suggestions. Please let me know if there are any issues remaining. I will keep #70 open until all the review process was completed.

Have a nice day!

> The source code comment in the time series extraction example "Image or ImageCollection (mean composite)" is not so clear, I'd suggest to describe what you actually do in the lines below (select dataset, reproject, filter by date and variable).

You're right, I made changes in the paper to solve this issue.

> The paper mentions that rgee is available from CRAN but it is not (yet).

Although I made the changes before. It seems I not compiled the paper correctly. I remove this part of the paper.
We have some problems to publish the package in CRAN (It was because rgee did not respect all CRAN policy since many functions used to manipulate the .Renviron without consulting users). We decided to wait to have the JOSS decision before resubmitting the package.

> For the region argument in ee_as_stars() and ee_as_raster(), the documentation states that "if not specified, image bounds will be taken". However, in my tests, I get "Error in ee_image_local(image = image, region = region, dsn = dsn, via = via, : argument "region" is missing, with no default" when calling these functions without region argument.

Thanks, we fix this, now the following code will work:

```r
library(rgee)
library(stars)
ee_Initialize()

ee$Image("LANDSAT/LC08/C01/T1_SR/LC08_038029_20180810") %>%
  ee$Image$select(c("B4", "B3", "B2")) %>%
  ee$Image$divide(10000) %>%
  ee_as_stars(scale = 1000) -> l8_img
plot(l8_img, rgb = c(1,2,3))
```
>  I've found that calling ee_Initialize(drive = TRUE, gcs = TRUE) sometimes constantly asks for a selection of the mail address (though there is only one listed). If selecting 1, it asks again whereas if selecting 0, the package opens the browser for authorization, and after allowing everything, it asks again. Am I doing something wrong? I don't know any Google API details but if possible, I think it would be nice to make the Google drive / cloud storage authorization as simple as when calling ee_Initialize() for GEE only (which works well).

This was a big mistake. Thank you very much for pointing it out. It was fixed, please let us know if the problem persists.

> It is clear that function examples in the documentation are marked as dontrun, because they need a GEE account. However, I would have enjoyed to run R CMD check with argument --run-dontrun with my own GEE acount, and also run the testthat tests. I was not successful because it seems that the GEE user is hard-coded in the tests and some examples (with Google drive or cloud storage) do not work in a non-interactive session. Do you see any way to make it easier for others to run all tests / examples automatically?

We completely change the test files. Now everyone can run tests smoothly. If you try to run a test without GCS credentials, the specific test will simply be skipped. All the necessary parameters to run tests are set in testthat.R file. You can also run R CMD check with the argument --run-dontrun with no problems. drive credentials are necessary.

> R CMD check --as-cran reports some documentation issues of print.ee.computedobject.ComputedObject(): "Found the following apparent S3 methods exported but not registered: print.ee.computedobject.ComputedObject"

Thanks, it was fixed.
