+++ R/sf_as_ee.R
@@ -26,8 +26,8 @@
 #' @param bucket Character. Name of the bucket (GCS) to save intermediate files
 #' (ignore if \code{via} is not defined as "gcs_to_asset").
 #' @param command_line_tool_path Character. Path to the Earth Engine command line
-#' tool. If NULL, rgee assumes is saved in the same path that your Python
-#' environment (ignore if \code{via} is not defined as "gcs_to_asset").
+#' tool (CLT). If NULL, rgee assumes that CLT is set in the system PATH.
+#' (ignore if \code{via} is not defined as "gcs_to_asset").
 #' @param overwrite A boolean argument which indicates indicating
 #' whether "filename" should be overwritten. Ignore if \code{via} argument
 #' is "getInfo". By default TRUE.
@@ -227,7 +227,7 @@ sf_as_ee <- function(x,
     }
 
     if (is.null(command_line_tool_path)) {
-      command_line_tool_path <- dirname(Sys.getenv("EARTHENGINE_PYTHON"))
+      command_line_tool_path <- ""
     }
 
     shp_dir <- sprintf("%s.shp", tempfile())
