+++ man/gcs_to_ee_image.Rd
@@ -33,8 +33,8 @@ it exists.}
 of the EE object.}
 
 \item{command_line_tool_path}{Character. Path to the Earth Engine command line
-tool. If NULL, rgee assumes is saved in the same path that your Python
-environment.}
+tool (CLT). If NULL, rgee assumes that CLT is set in the system PATH.
+(ignore if \code{via} is not defined as "gcs_to_asset").}
 
 \item{start_time}{Character. Sets the start time property to a number
 or date.}
