### 2020-07-05 15847d2
@gena comments
- Fix a bug in GH actions
