### 2020-07-05 r-spatial/rgee#72 @csaybar
Hi @gena thank you so much for taking your time in review rgee. With respect to your question, we use the `system.file` function to create reproducible examples, this function helps to retrieve the full file names of files in R packages installed. When a package is installed, everything in [inst/](https://github.com/r-spatial/rgee/tree/master/inst) folder is copied into the top-level package directory (see R code below). We take advance of this and save a toy shapefile [inst/shp](https://github.com/r-spatial/rgee/tree/master/inst/shp) to create reproducible examples and run unit tests.

``` r
list.files(system.file("", package = "rgee"))
#>  [1] "CITATION"    "DESCRIPTION" "help"        "html"        "INDEX"
#>  [6] "Meta"        "NAMESPACE"   "NEWS.md"     "python"      "R"
#> [11] "shp"         "tests"       "WORDLIST"
```

I think the problem here is that sf is maybe not installed (we use its inst/ files in example 2), we add extra info in the README.rd file for users to take into account that to run example 2 sf must be installed. Please let me know if the problem persists. I will keep #72 open until all the review process was completed.

Have a good Sunday!
