### 2020-07-05 fa57b70
mapview problem in MacOS GH actions
2020-07-05T13:02:07.9910820Z   dlopen(/Users/runner/work/_temp/Library/systemfonts/libs/systemfonts.so, 6): Library not loaded: /opt/X11/lib/libfreetype.6.dylib
2020-07-05T13:02:07.9911150Z   Referenced from: /Users/runner/work/_temp/Library/systemfonts/libs/systemfonts.so
2020-07-05T13:02:07.9911350Z   Reason: image not found
2020-07-05T13:02:07.9911510Z Backtrace:
2020-07-05T13:02:07.9911720Z   1. mapview::mapview(nc)
2020-07-05T13:02:07.9911910Z   5. mapview::mapview(...)
2020-07-05T13:02:07.9912110Z   7. mapview::mapView(...)
2020-07-05T13:02:07.9912310Z
[…]
