### 2020-07-05 r-spatial/rgee#72 @gena
You're right, I could reproduce the example perfectly, thanks for clarifying this!
