+++ DESCRIPTION
@@ -75,7 +75,7 @@ Authors@R:
              comment = "Appel reviewed the package for JOSS, see 
                         https://github.com/openjournals/joss-reviews/issues/2272"))
 Description: Earth Engine <https://earthengine.google.com/> client library for R. All of the 'Earth Engine' API classes, 
-    modules, and functions are made available. Additional functions implemented include importing (exporting) of R Earth Engine 
+    modules, and functions are made available. Additional functions implemented include importing (exporting) of Earth Engine 
     spatial objects, extraction of time series, interactive map display, assets management interface, and metadata display.
     See <https://r-spatial.github.io/rgee/> for further details.
 License: Apache License (>= 2.0)
