### 2020-07-06 #2272 @hugoledoux
@csaybar has https://github.com/r-spatial/rgee/issues/72 been completely solved? It seems yes but you left it open. It seems all reviewers are happy now, and if this issue is closed then we can move towards acceptance and the final steps.
