### 2020-07-06 r-spatial/rgee#72 @csaybar
Hi @gena to move towards rgee acceptance (openjournals/joss-reviews#2272), could we close the issue now?
