### 2020-07-06 9342626
Fix a bug in ee_get_date_ic
