### 2020-07-09 ff4a3f2
Map$addLayer now display a legend!
Earth Engine Python API updated to 0.1.227
