### 2020-07-09 163b955
changes in test-map.R
