### 2020-07-10 7f7cda6
Minor bug in Map$addLayer.
Map$addLayer creates EarthEngineMap objects but not display them.
