### 2020-07-11 0d67157
Minor bug in Image PixelType
