### 2020-07-13 #2272 @csaybar
Hi @hugoledoux sorry for the last reply. We are waiting for the response of @appelmar (r-spatial/rgee#70)!.
