### 2020-07-14 r-spatial/rgee#70 @appelmar
Great, thanks for addressing the issues, well done!
