### 2020-07-14 13ee3a4
ee_get and Map$addLayer added
