+++ R/ee_Date.R
@@ -76,147 +76,3 @@ eedate_to_rdate <- function(ee_date, timestamp = FALSE) {
     as.POSIXct(x = date_numeric/1000, origin = "1970-01-01", tz = "GMT")
   }
 }
-
-#' Get the date of a EE Image
-#'
-#' @param x ee$Image or ee$ImageCollection object
-#' @param time_end Logical. If TRUE, the \code{system:time_end} property is
-#' also returned. See details.
-#'
-#' @details
-#' \code{system:time_start} set the start period of data acquisition while
-#' \code{system:time_end} does the same for the end period. See the
-#' \href{https://developers.google.com/earth-engine/glossary}{Earth Engine glossary}
-#' for getting more information.
-#'
-#' @return An List object with the elements: id, time_start and time_end
-#' (if the \code{time_end} argument is TRUE).
-#' @family date functions
-#' @examples
-#' \dontrun{
-#' library(rgee)
-#' ee_Initialize()
-#'
-#' l8 <- ee$Image('LANDSAT/LC08/C01/T1_TOA/LC08_044034_20140318')
-#' ee_get_date_img(l8)
-#' srtm <- ee$Image('CGIAR/SRTM90_V4')
-#' ee_get_date_img(srtm, time_end = TRUE)
-#' }
-#' @export
-ee_get_date_img <- function(x, time_end = FALSE) {
-  time_start <- tryCatch(
-    expr = eedate_to_rdate(x$get("system:time_start")),
-    error = function(e) NA
-  )
-  if (isTRUE(time_end)) {
-    time_end <- tryCatch(
-      expr = eedate_to_rdate(x$get("system:time_end")),
-      error = function(e) NA
-    )
-  } else {
-    time_end <- NULL
-  }
-  # get id of the image
-  image_id <- x$get("system:id")$getInfo()
-  if (is.null(image_id)) {
-    image_id <- "no_id"
-  }
-  if (is.null(time_end)) {
-    list(image_id = image_id, time_start = time_start)
-  } else {
-    list(image_id = image_id, time_start = time_start, time_end = time_end)
-  }
-}
-
-#' Get the date of a EE ImageCollection
-#'
-#' @param x ee$ImageCollection object
-#' @param time_end Logical. If TRUE, the \code{system:time_end} property is
-#' also returned. See details.
-#'
-#' @details
-#' \code{system:time_start} set the start period of data acquisition while
-#' \code{system:time_end} does the same for the end period. See the
-#' \href{https://developers.google.com/earth-engine/glossary}{Earth Engine glossary}
-#' for getting more information.
-#'
-#' @return A data.frame with the columns: \code{id} (ID of the image),
-#' \code{time_start}, and \code{time_end} (If the argument \code{time_end} is
-#' set as TRUE). The number of rows depends of the number of images
-#' (\code{ee$ImageCollection$size}).
-#'
-#' @family date functions
-#'
-#' @examples
-#' \dontrun{
-#' library(rgee)
-#' library(sf)
-#' ee_Initialize()
-#'
-#' nc <- st_read(system.file("shape/nc.shp", package = "sf")) %>%
-#'   st_transform(4326) %>%
-#'   sf_as_ee()
-#'
-#' ee_s2 <- ee$ImageCollection("COPERNICUS/S2")$
-#'   filterDate("2016-01-01", "2016-01-31")$
-#'   filterBounds(nc)
-#'
-#' ee_get_date_ic(ee_s2)
-#' }
-#' @export
-ee_get_date_ic <- function(x, time_end = FALSE) {
-  # Call Python module
[…]
