+++ pkgdown/_pkgdown.yml
@@ -78,7 +78,6 @@ reference:
       - EarthEngineMap-class
       - +,EarthEngineMap,EarthEngineMap-method
       - +,mapview,EarthEngineMap-method
-      - print,EarthEngineMap-method
       - show,EarthEngineMap-method
       - mapview-class
       - +,EarthEngineMap,mapview-method
@@ -146,6 +145,7 @@ reference:
       - ee_print
       - print.ee.computedobject.ComputedObject
       - ee_monitoring
+      - ee_get
 
   - title: "Utils"
     contents:
