### 2020-07-15 #2272 @hugoledoux
@csaybar there's a glitch with your PDF, see this:

![CleanShot 2020-07-15 at 08 03 01@2x](https://user-images.githubusercontent.com/1546518/87509593-2bb3da80-c672-11ea-91aa-8de832a7f860.png)

Something goes wrong with a function name with "$ and it doesn't escape, I think you should modify at l.45
```
("$getInfo")
```
=>
```
`($getInfo)`
```
