### 2020-07-15 #2272 @hugoledoux
also, shouldn't the reference to mapview be the package/website? I tried to find the paper, but couldn't in 2min, while the website is complete, informative, etc. If you want to stick to the paper at EGU, please provide a DOI
