### 2020-07-15 #2272 @hugoledoux
@csaybar the 3 reviewers recommend acceptance, and I'm happy to say that we're reaching the end of the review!

At this point could you:

- [ ] Fix the 2 issues I pointed above
- [ ] Archive the reviewed software in Zenodo or a similar service (e.g. figshare, an institutional repository)
- [ ] Check the archival deposit (e.g., in Zenodo) has the correct metadata, this includes the title (should match the paper title) and author list (make sure the list is correct and people who only made a small fix are not on it); you may also add the authors' ORCID.

Please list the DOI of the archived version here.
