### 2020-07-15 9c0c43b
fix paper issues
