+++ paper/paper.bib
@@ -9,14 +9,6 @@
   publisher={Elsevier}
 }
 
-@inproceedings{appelhans2016mapview,
-  title={mapview-Interactive viewing of spatial data in R},
-  author={Appelhans, Tim and Detsch, Florian and Reudenbach, Cristoph and Woellauer, Stefan},
-  booktitle={EGU General Assembly Conference Abstracts},
-  volume={18},
-  year={2016}
-}
-
 @article{wickham2011ggplot2,
   title={ggplot2},
   author={Wickham, Hadley},
@@ -30,8 +22,17 @@
 }
 
 @Manual{reticulate,
-  author = {JJ Allaire, Kevin Ushey, Yuan Tang and Dirk Eddelbuettel},
-  title = {reticulate: R Interface to Python},
-  year = {2017},
-  url = {https://github.com/rstudio/reticulate},
+  title = {reticulate: Interface to 'Python'},
+  author = {Kevin Ushey and JJ Allaire and Yuan Tang},
+  year = {2020},
+  note = {R package version 1.16},
+  url = {https://CRAN.R-project.org/package=reticulate},
+}
+
+@Manual{mapview,
+    title = {mapview: Interactive Viewing of Spatial Data in R},
+    author = {Tim Appelhans and Florian Detsch and Christoph Reudenbach and Stefan Woellauer},
+    year = {2020},
+    note = {R package version 2.8.1},
+    url = {https://github.com/r-spatial/mapview},
 }
