### 2020-07-15 #2272 @csaybar
Dear @hugoledoux thanks for the fast response. I've made the requested changes.
- I changed ("$getInfo") by ("\$getInfo") the problem will be fixed now. With respect to the citation to mapview we now refer to the R package manual instead of the publication in EGU.
- rgee [v.0.5.0](https://github.com/r-spatial/rgee/releases/tag/v.0.5.0).
- rgee is archived in zenodo [here](https://zenodo.org/record/3946739).
