### 2020-07-16 #2272 @hugoledoux
@csaybar I am confused, you have a v1.0.0 release, but also you made one for v.0.5.0? Your releases have 2 options for one repository? This is new to me, and rather confusing... Can you shed some light on why that is?
