### 2020-07-16 #2272 @csaybar
Hi Hugo, after submitting rgee version 0.5.0 to JOSS I was making some additional changes that were reflected in new versions of rgee. Version 1.0.0 has all the changes made to pass the issues pointed out by the reviewers. However, I was not sure which version should be used if version 0.5.0 or 1.0.0. Sorry if this is a bit confusing.
