### 2020-07-16 #2272 @hugoledoux
ah right, then if you released v1.0.0 based on all reviews than it should be v.1.0.0 for JOSS, this is a nice outcome too.

Could you make a new Zenodo release (pointing to v1.0.0) and list the DOI here please?
