### 2020-07-16 #2272 @csaybar
@hugoledoux  I've made the requested changes.

- Archived rgee v.1.0.0 in Zenodo [here](https://zenodo.org/record/3945409).
- rgee v.1.0.0 Zenodo citation: Cesar Aybar, Qiusheng Wu, Lesly Bautista, Roy Yali, & Antony Barja. (2020, July 14). rgee: An R package for interacting with Google Earth Engine (Version v.1.0.0). Zenodo. http://doi.org/10.5281/zenodo.3945409
