+++ paper/paper.bib
@@ -10,7 +10,7 @@
 }
 
 @article{wickham2011ggplot2,
-  title={ggplot2},
+  title={{ggplot2}},
   author={Wickham, Hadley},
   journal={Wiley Interdisciplinary Reviews: Computational Statistics},
   volume={3},
@@ -22,7 +22,7 @@
 }
 
 @Manual{reticulate,
-  title = {reticulate: Interface to 'Python'},
+  title = {{reticulate}: Interface to 'Python'},
   author = {Kevin Ushey and JJ Allaire and Yuan Tang},
   year = {2020},
   note = {R package version 1.16},
@@ -30,7 +30,7 @@
 }
 
 @Manual{mapview,
-    title = {mapview: Interactive Viewing of Spatial Data in R},
+    title = {{mapview}: Interactive Viewing of Spatial Data in R},
     author = {Tim Appelhans and Florian Detsch and Christoph Reudenbach and Stefan Woellauer},
     year = {2020},
     note = {R package version 2.8.1},
