### 2020-07-16 #2272 @hugoledoux
two small things:

1. https://github.com/r-spatial/rgee/pull/74

2. there is something strange with your zenodo, if I go to the v1.0.0 I get that warning, which redirects me to v0.5.0, not sure why... could you look into this?

![CleanShot 2020-07-16 at 08 34 27@2x](https://user-images.githubusercontent.com/1546518/87635449-5cae1100-c73f-11ea-89a8-6acc7069b3b1.png)
