### 2020-07-16 #2272 @csaybar
@hugoledoux Sorry for the inconvenience, but I had to create a new archived in Zenodo to solve the issue.

- New archived rgee v.1.0.0 in Zenodo [here](https://zenodo.org/record/3947571).
- rgee v.1.0.0 Zenodo citation: Cesar Aybar, Qiusheng Wu, Lesly Bautista, Roy Yali, & Antony Barja. (2020, July 16). rgee: An R package for interacting with Google Earth Engine (Version v.1.0.0). Zenodo. http://doi.org/10.5281/zenodo.3947571
- Thanks for the PR, I added it to the master branch.
